chore: get only tokens worth minimum .1 eth and sort by updated at to make random
